fix: add new component usePolicy
